+++ spec/lib/paypal-api/client_spec.rb
@@ -72,6 +72,7 @@ RSpec.describe PaypalAPI::Client do
       invoice_templates: PaypalAPI::InvoiceTemplates,
       invoices: PaypalAPI::Invoices,
       orders: PaypalAPI::Orders,
+      partner_referrals: PaypalAPI::PartnerReferrals,
       payment_tokens: PaypalAPI::PaymentTokens,
       payout_items: PaypalAPI::PayoutItems,
       payouts: PaypalAPI::Payouts,